o Basic zooming support. Still a few bugs but the basics work. o Adding a new window showing the list of internal exceptions, IDEA style. o Adding a global exception handler for all exceptions. Seems to only handle exceptions from actions, not tasks.

+++ infero-gui/src/main/java/infero/util/NanoSeconds.java
@@ -25,7 +25,7 @@ public class NanoSeconds {
         return new NanoSeconds(microSeconds * 1000);
     }
 
-    public static NanoSeconds nanoSeconds(int nanoSeconds) {
+    public static NanoSeconds nanoSeconds(long nanoSeconds) {
         return new NanoSeconds(nanoSeconds);
     }
